# src/user/dto/create-user.dto.ts
import { IsNotEmpty, IsString } from 'class-validator';

export class CreateUserDto {
  @IsNotEmpty()
  @IsString()
  login: string;

  @IsNotEmpty()
  @IsString()
  password: string;
}

export interface ICreateUserDto {
  login: string;
  password: string;
}